Remove Connection type from Gramophone.Database
The purpose of the Connection type was to allow a reference
to an open database to be stored outside the model without
exposing the underlying database. Since the Database module no
longer exposes references to open databases (All exported
functions close the database before returning), the Connection
type was no longer necessary.