Replace the DBIO monad with a DBT monad transformer
The DBT Monad transformer is used for functions that need to
access the database. It is an instance of the MonadDB class
which is what most functions outside the Database module will
use.

DBT is *not* in instance of MonadTrans--it can only be used to
transform monads which are MonadIO instances. It would probably
have been possible to make it a MonadTrans instance, by removing
the MonadIO requirement, and that would have also simplified
some of the code. However, since you can't actually do anything
with a database connection without being in IO, that would have
forced all functions that required a MonadDB to also require
MonadIO.