find(Artists|Albums) expect (Album|Artist)Name argument, not Text
When looking up artist an albums by name, need to pass AlbumName
or ArtistName types as the argument, rather than simply Text as
before.

I'm not sure if the new or old way is better, as far as API
clarity goes, but doing it this way makes the process of adding
new recordings simpler.