Add some extra automatically-derived instances
The Id type, for id (primary key) columns in the database, is now
an instance of Read and Eq, which are both required so that
ids can be used in URLs in the web GUI. Also, they are
automatically derived and (Eq in particular) useful in general.

Also added Ord instances to TrackCount and TrackNumber, with
obvious usefulness.