Allow conversion of Id to/from text
RecordingID, AlbumID and ArtistID can now be converted to/from
hexadecimal text strings. This is necessary to allow them to
be specified in routes by the Web GUI.

An alternative would have been to expose the constructor and
thereby allow conversion to and from Integers, but I think
supplying this converter to and from strings is better because
it mostly preserves the opaqueness of the types.

Of course, I could just use read and show but this approach has
the advantage of compactness and Haskell-independence.